convert slashCommands archive files to ts
