--- app/slashcommands-archiveroom/client/client.js
@@ -1,7 +1,0 @@
-import { slashCommands } from '../../utils';
-
-slashCommands.add('archive', null, {
-	description: 'Archive',
-	params: '#channel',
-	permission: 'archive-room',
-});
